Approximate conic by Bézier curves
Choosing the control points currently means solving a fourth degree
polynomial, and it appears to be very hard to choose the right solution.
It's not even clear to me right now whether there IS a unique right
solution.  So that choice of segment needs to be handled before we can
investigate adaptive refinement.