Use heuristics to choose Bézier candidate
Looking at an example using the previous SVG export facility, it looks as
though there often are three candidates for a given segment.  Two of them
can be described as follows: one of the control points is very near to its
endpoint, while the other control point is very far from its endpoint, so
much so that it lies approximately in the tangent direction of the other
endpoint.  The third result looks more balanced.  Since the balanced result
feels most symmetric and least degenerate, there now is a heuristic in place
to prefer such balanced setups.