Disprove of concept: refinement via t=.5 point doesn't work
This tries out one possible approach for recursive refinement.  Once the
Bézier control points have been computed, the parameter t = 0.5 is plugged
into the equation to obtain a point along the line.  The gradient at that
point is computed and used to judge how far the curve is from the true
conic.

This has the effect of making the conic really sluggish, but fails to result
in significantly better approximations.  Apparently the fact that this
single point is reasonably close is no reliable indication that the curve as
a whole is close.  And the computational cost for the whole approach appear
to be considerable.  There are also some visible artefacts, questioning the
value of previous heuristics and the overall goal of using Bézier curves as
a means to improve quality.