Switch from cubic to quadratic Bézier splines
These are WAY cheaper to compute, so we can use many more of them.
Compared to piecewise linear rendering, we still get consistent tangent
directions, so we don't see corners in conic sections.

Also fixed some bugs in the code projecting the point onto the conic.  That
MIGHT have been accountable for the fact that cubic splines became so slow,
but on the whole this approach here feels fare more appropriate.

There are still some issues, and one of them is captured by the newly added
example.