Vertically clip to canvas one step earlier
Now we only connect points if they are actually part of an on-canvas arc
which definitely will get rendered.  This gives us better control about
whether we may eventually close the path or not.  We will end up with
spurious start points, which are not connected to anything, but that doesn't
hurt.